+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.10.4
 
 ### New Features
+- Show dialog to confirm users action when creating a server without any instances
 
 ### Fixes
 - Issue with CurseForge packs using forge-recommended or forge-latest failing to install